+++ lib/services/object_detection/object_detection_service.dart
@@ -4,18 +4,18 @@ import "dart:typed_data";
 import "package:logging/logging.dart";
 import "package:photos/services/object_detection/models/predictions.dart";
 import 'package:photos/services/object_detection/models/recognition.dart';
-import "package:photos/services/object_detection/tflite/classifier.dart";
+import 'package:photos/services/object_detection/tflite/cocossd_classifier.dart';
+import "package:photos/services/object_detection/tflite/mobilenet_classifier.dart";
 import "package:photos/services/object_detection/utils/isolate_utils.dart";
 
 class ObjectDetectionService {
-  static const scoreThreshold = 0.6;
+  static const scoreThreshold = 0.5;
 
   final _logger = Logger("ObjectDetectionService");
 
-  /// Instance of [ObjectClassifier]
-  late ObjectClassifier _classifier;
+  late CocoSSDClassifier _objectClassifier;
+  late MobileNetClassifier _mobileNetClassifier;
 
-  /// Instance of [IsolateUtils]
   late IsolateUtils _isolateUtils;
 
   ObjectDetectionService._privateConstructor();
@@ -23,7 +23,8 @@ class ObjectDetectionService {
   Future<void> init() async {
     _isolateUtils = IsolateUtils();
     await _isolateUtils.start();
-    _classifier = ObjectClassifier();
+    _objectClassifier = CocoSSDClassifier();
+    _mobileNetClassifier = MobileNetClassifier();
   }
 
   static ObjectDetectionService instance =
@@ -31,23 +32,50 @@ class ObjectDetectionService {
 
   Future<List<String>> predict(Uint8List bytes) async {
     try {
-      final isolateData = IsolateData(
-        bytes,
-        _classifier.interpreter.address,
-        _classifier.labels,
-      );
-      final predictions = await _inference(isolateData);
-      final Set<String> results = {};
-      for (final Recognition result in predictions.recognitions) {
-        if (result.score > scoreThreshold) {
-          results.add(result.label);
-        }
-      }
+      final results = <String>{};
+      final objectResults = await _getObjects(bytes);
+      results.addAll(objectResults);
+      final mobileNetResults = await _getMobileNetResults(bytes);
+      results.addAll(mobileNetResults);
       return results.toList();
     } catch (e, s) {
       _logger.severe(e, s);
       rethrow;
     }
+  }
+
+  Future<List<String>> _getObjects(Uint8List bytes) async {
+    final isolateData = IsolateData(
+      bytes,
+      _objectClassifier.interpreter.address,
+      _objectClassifier.labels,
+      ClassifierType.cocossd,
+    );
+    final predictions = await _inference(isolateData);
+    final Set<String> results = {};
+    for (final Recognition result in predictions.recognitions) {
+      if (result.score > scoreThreshold) {
+        results.add(result.label);
+      }
+    }
+    return results.toList();
+  }
+
+  Future<List<String>> _getMobileNetResults(Uint8List bytes) async {
+    final isolateData = IsolateData(
+      bytes,
+      _mobileNetClassifier.interpreter.address,
+      _mobileNetClassifier.labels,
+      ClassifierType.mobilenet,
+    );
+    final predictions = await _inference(isolateData);
+    final Set<String> results = {};
+    for (final Recognition result in predictions.recognitions) {
+      if (result.score > scoreThreshold) {
+        results.add(result.label);
+      }
+    }
+    return results.toList();
   }
 
   /// Runs inference in another isolate
