+++ lib/services/object_detection/utils/isolate_utils.dart
@@ -2,7 +2,8 @@ import 'dart:isolate';
 import "dart:typed_data";
 
 import 'package:image/image.dart' as imgLib;
-import "package:photos/services/object_detection/tflite/classifier.dart";
+import 'package:photos/services/object_detection/tflite/cocossd_classifier.dart';
+import "package:photos/services/object_detection/tflite/mobilenet_classifier.dart";
 import 'package:tflite_flutter/tflite_flutter.dart';
 
 /// Manages separate Isolate instance for inference
@@ -29,10 +30,17 @@ class IsolateUtils {
     sendPort.send(port.sendPort);
 
     await for (final IsolateData isolateData in port) {
-      final classifier = ObjectClassifier(
-        interpreter: Interpreter.fromAddress(isolateData.interpreterAddress),
-        labels: isolateData.labels,
-      );
+      final classifier = isolateData.type == ClassifierType.cocossd
+          ? CocoSSDClassifier(
+              interpreter:
+                  Interpreter.fromAddress(isolateData.interpreterAddress),
+              labels: isolateData.labels,
+            )
+          : MobileNetClassifier(
+              interpreter:
+                  Interpreter.fromAddress(isolateData.interpreterAddress),
+              labels: isolateData.labels,
+            );
       final image = imgLib.decodeImage(isolateData.input);
       final results = classifier.predict(image!);
       isolateData.responsePort.send(results);
@@ -45,11 +53,18 @@ class IsolateData {
   Uint8List input;
   int interpreterAddress;
   List<String> labels;
+  ClassifierType type;
   late SendPort responsePort;
 
   IsolateData(
     this.input,
     this.interpreterAddress,
     this.labels,
+    this.type,
   );
+}
+
+enum ClassifierType {
+  cocossd,
+  mobilenet,
 }
